% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{run_table_example}
\alias{run_table_example}
\title{Run Shiny React Table example}
\usage{
run_table_example(...)
}
\arguments{
\item{...}{Parameters for \code{\link[shiny]{runApp}}.}
}
\description{
An example page showing a few use cases and error handling.
}
