% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-react.r
\name{table_react}
\alias{table_react}
\alias{table_react_select}
\alias{table_react_update}
\title{Shiny React Table}
\usage{
table_react(id)

table_react_select(session, id, select_group = NULL)

table_react_update(session, id, data = NULL, select_group = NULL,
  page = NULL, page_size = NULL)
}
\arguments{
\item{id}{Component ID for updating}

\item{session}{shiny session}

\item{select_group}{Replace the selection. A \code{\link[base]{integer}} vector containing.
\code{\link[base]{NA}} for unselected, and numbers for selected rows.}

\item{data}{Data for the table. Should be convertible to a \code{\link[base]{data.frame}}.}

\item{page}{Select a page for pagination.}

\item{page_size}{Set a page size for pagination.}
}
\description{
A fast table with multiple selection.
\code{input$my_table_id} has the same structure as the parameter \code{select_group}.
}
